requirements.txt
numpy==1.23.0
h5py==3.6.0
nilearn==0.9.2
tqdm==4.64.1
darts==0.16.0
pandas==1.3.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.15
